+++ Data/Retail/ExportedWidgetEvents.lua
@@ -24,12 +24,13 @@ end
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-10-31 20-52-17 ]] --
+-- [[ Exported at 2023-11-03 10-57-23 ]] --
 tasks = {
     {N, 4729, 1393, 4240492, addon.L["Community Feast: Starting Soon"]}, -- Community Feast: Starting Soon
     {N, 4731, 1393, 4240492, addon.L["Community Feast: Active"]}, -- Community Feast: Active
     {N, 4987, 1454, 236387, addon.L["Time Rift: Active"]}, -- Time Rift: Active
     {N, 4992, 1454, 236387, addon.L["Time Rift: Starting Soon"]}, -- Time Rift: Starting Soon
+    {N, 5157, 1552, 3939983, addon.L["Superbloom: Next"]}, -- Superbloom: Next
     {N, 5323, 1552, 3939983, addon.L["Superbloom: Active"]}, -- Superbloom: Active
 };
 
@@ -40,6 +41,7 @@ function exportedWidgetEvents.InjectDynamicOptions()
     KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", addon.L["Time Rifts"], {4987}, addon.L["Time Rift: Active"]);
     KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", addon.L["Time Rifts"], {4992}, addon.L["Time Rift: Starting Soon"]);
     KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Time_Rifts", { 4987, 4992 });
+    KrowiAF_RegisterEventOptions("Widget", "Superbloom", addon.L["Superbloom"], {5157}, addon.L["Superbloom: Next"]);
     KrowiAF_RegisterEventOptions("Widget", "Superbloom", addon.L["Superbloom"], {5323}, addon.L["Superbloom: Active"]);
-    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Superbloom", { 5323 });
+    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Superbloom", { 5157, 5323 });
 end